Flask message modification.

+++ src/controllers/accountController.py
@@ -30,10 +30,9 @@ def setting():
 #IDOR-3 - END
 
 #StoredXSS-2 - START
-"""Status: Vulnerable"""
+"""Status: Fixed"""
 #Description: CWE-79: Cross-site Scripting -> https://cwe.mitre.org/data/definitions/79.html
 def update_user():
-    #referrer_url = request.referrer
     if request.method == 'POST':
         try:
             validate_csrf(request.form.get('csrf_token'))
@@ -47,53 +46,58 @@ def update_user():
             first_name = request.form.get("edit_fn")
             last_name = request.form.get("edit_ln")
             password = request.form.get("edit_password")
+
             user = User.query.filter_by(id=id).first()
             current_username = user.username
             current_email = user.email
             if current_username != username:
-                #User input is not beeing validated in any way.
+                #Function compares user input against allowed pattern.
+                input_validation(username, "Username")
                 check_if_exists('username', username, 'Username')
                 user.username = username
             if current_email != email:
-                #User input is not beeing validated in any way.
+                #Function compares user input against allowed pattern.
+                email_validation(email)
                 check_if_exists('email', email, 'Email')
                 user.email = email 
-            #User input is not beeing validated in any way.     
             user.first_name = first_name
-            #User input is not beeing validated in any way.
+            #Function compares user input against allowed pattern.
+            input_validation(user.first_name, "First Name")
             user.last_name = last_name
+            #Function compares user input against allowed pattern.
+            input_validation(user.last_name, "Last Name")
             if password == '':
                 pass
             else:
                 #WeakPasswordRequirements-3 - START
-                """Status: Vulnerable"""
+                """Status: Fixed"""
                 #Description: CWE-521: Weak Password Requirements -> https://cwe.mitre.org/data/definitions/521.html
-                #There is no check of length and complexity of a password.
+                check_for_password_complexity(password)
                 #WeakPasswordRequirements-3 - END
                 #CompleteOmissionOfHashFunction-1 - START
                 #CompleteOmissionOfHashFunction-1 - END
                 #WeakHashFunction-1 - START
-                """Status: Vulnerable"""
-                #Description: CWE-327: Use of a Broken or Risky Cryptographic Algorithm -> https://cwe.mitre.org/data/definitions/327.html
-                password = md5(password.encode()).hexdigest()
                 #WeakHashFunction-1 - END
                 #WeakHashFunctionWithSalt-1 - START
-                #WeakHashFunctionWithSalt-1 - END  
+                """Status: Vulnerable"""
+                #Description: CWE-327: Use of a Broken or Risky Cryptographic Algorithm -> https://cwe.mitre.org/data/definitions/327.html
+                password = md5_crypt.using(salt_size=8).hash(password)
+                #WeakHashFunctionWithSalt-1 - END 
                 user.password = password
             db.session.commit()
-            log_config.logger.info("User with username %s was sucessfully updated." % bleach.clean(username), extra={'ip_address': request.remote_addr})
+            log_config.logger.info("User with username %s was sucessfully updated." %  bleach.clean(username), extra={'ip_address': request.remote_addr})
             flash("User has been updated.", 'success')
-            return redirect(request.referrer)  
+            return redirect(request.referrer)
         except ValidationError:
             log_config.logger.error("User was not updated. Missing or invalid CSRF token.", extra={'ip_address': request.remote_addr})
             abort(400)
         except ValueError:
             return redirect(request.referrer)
         except Exception as e:
-            log_config.logger.error("User was not sucessfully updated. Exception: %s" % e, extra={'ip_address': request.remote_addr})
+            log_config.logger.error("User was not successfully updated. Exception: %s" % e, extra={'ip_address': request.remote_addr})
             flash("Error occured, try again.", 'danger')
-            return redirect(request.referrer)
-    return redirect(request.referrer)
+            return redirect(request.referrer)     
+    return redirect(request.referrer)  
 #StoredXSS-2 - END
 
 
@@ -107,6 +111,9 @@ def delete_user():
             user = User.query.filter_by(id=id).first()
             if user is not None:
                 #SensitiveDatawithinCookie-2 - START
+                """Status: Fixed"""
+                #Description: CWE-315: Cleartext Storage of Sensitive Information in a Cookie -> https://cwe.mitre.org/data/definitions/315.html
+                """Since 'role' is not a part of a session, there is no need to do anything at this point."""
                 #SensitiveDatawithinCookie-2 - END
                 session.pop('cart')
                 session.pop('total')
@@ -135,8 +142,29 @@ def orders():
 
 
 #MaliciousFileUpload-1 - START  
-"""Status: Vulnerable"""
+"""Status: Fixed"""
 #Description: CWE-434: Unrestricted Upload of File with Dangerous Type -> https://cwe.mitre.org/data/definitions/434.html
+ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif', 'svg'}
+ALLOWED_MIME = {'image/png','image/jpg','image/jpeg', 'image/svg+xml'}
+
+def check_extension(filename):
+    """Function returns True if extension is all right, and False if otherwise."""
+    filename_split = filename.split('.')
+    if len(filename_split) != 2:
+        return False
+    else:
+        extension = filename_split[-1]
+        if extension in ALLOWED_EXTENSIONS:
+            return True
+        else:
+            return False
+
+def check_mime(picture):
+    if mimetypes.guess_type(picture.filename)[0] in ALLOWED_MIME:
+        return True
+    else:
+        return False
+
 def upload_picture():
     if request.method == 'POST':
         try:
@@ -146,6 +174,7 @@ def upload_picture():
             # check if the post request has the file part
             if 'profile_picture' not in request.files:
                 flash('No file part', 'danger')
+                print('No file part')
                 return redirect(request.referrer)
             picture = request.files['profile_picture']
             #print("Filename: ", picture.filename)
@@ -153,26 +182,37 @@ def upload_picture():
             # empty file without a filename.
             if picture.filename == '':
                 flash('No selected file', 'danger')
+                print('No selected file')
                 return redirect(request.referrer)
-            picture_uuid = str(uuid.uuid1())+ secure_filename(picture.filename)
-            if user.profile_picture:
-                    original_picture = secure_filename(user.profile_picture)
+            filename_sanitized = secure_filename(picture.filename)
+            print("Sanitized: ", filename_sanitized)
+            if picture and check_extension(filename_sanitized) and check_mime(picture):
+                file_to_store = str(uuid.uuid1())+filename_sanitized
+                if user.profile_picture:
+                    original_picture = user.profile_picture
                     print(original_picture)
                     path_to_original_picture = os.path.join(UPLOAD_FOLDER,original_picture)
                     print(path_to_original_picture)
                     os.remove(path_to_original_picture)
-            user.profile_picture = picture_uuid
-            db.session.commit()
-            upload_path = os.path.join(UPLOAD_FOLDER,picture_uuid)
-            picture.save(upload_path)
-            flash('Profile picture has been updated.', 'success')
-            log_config.logger.info("User %s successfully updated his profile picture." %  bleach.clean(user.username), extra={'ip_address': request.remote_addr})
-            return redirect(request.referrer)
+                user.profile_picture = file_to_store
+                db.session.commit()
+                upload_path = os.path.join(UPLOAD_FOLDER,file_to_store)
+                picture.save(upload_path)
+                flash('Profile picture has been updated.', 'success')
+                log_config.logger.info("User %s successfully updated his profile picture." % bleach.clean(user.username), extra={'ip_address': request.remote_addr})
+                return redirect(request.referrer)
+            else:
+                raise UnsupportedMediaType()
         except ValidationError:
             log_config.logger.error("User was not updated. Missing or invalid CSRF token.", extra={'ip_address': request.remote_addr})
             abort(400)
+        except RequestEntityTooLarge:
+            log_config.logger.error("User %s tried to upload file that exceeded allowed length limit." % bleach.clean(user.username), extra={'ip_address': request.remote_addr})
+            abort(413)
+        except UnsupportedMediaType:
+            abort(415)
         except Exception as e:
-            flash("Error occureed. Please try again.", 'danger')
+            flash("Error occureed. Please try again.",'danger')
             log_config.logger.error("User was not successfully updated. Exception: %s" % e, extra={'ip_address': request.remote_addr})
             return redirect(request.referrer) 
 #MaliciousFileUpload-1 - END
